server/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose
pillow
tensorflow
tensorflow-hub
opencv-python
numpy
mongoengine